refactor(checkout-button): CHECKOUT-3747 Group checkout button strategies in folders
