+++ src/checkout-buttons/strategies/braintree/index.ts
@@ -1,0 +1,2 @@
+export { default as BraintreePaypalButtonStrategy } from './braintree-paypal-button-strategy';
+export { BraintreePaypalButtonInitializeOptions } from './braintree-paypal-button-options';
