+++ src/checkout-buttons/strategies/index.ts
@@ -1,8 +1,7 @@
-export { default as BraintreePaypalButtonStrategy } from './braintree-paypal-button-strategy';
-export { BraintreePaypalButtonInitializeOptions } from './braintree-paypal-button-options';
+export { BraintreePaypalButtonStrategy, BraintreePaypalButtonInitializeOptions } from './braintree';
 export { GooglePayButtonStrategy, GooglePayButtonInitializeOptions } from './googlepay';
+export { MasterpassButtonStrategy } from './masterpass';
+export { PaypalButtonStrategy, PaypalButtonInitializeOptions } from './paypal';
+
 export { default as CheckoutButtonStrategy } from './checkout-button-strategy';
 export { CheckoutButtonMethodType } from './checkout-button-method-type';
-export { default as MasterpassButtonStrategy } from './masterpass-button-strategy';
-export { PaypalButtonInitializeOptions } from './paypal-button-options';
-export { default as PaypalButtonStrategy } from './paypal-button-strategy';
