+++ src/checkout-buttons/strategies/masterpass/index.ts
@@ -1,0 +1,1 @@
+export { default as MasterpassButtonStrategy } from './masterpass-button-strategy';
