+++ src/checkout-buttons/strategies/masterpass/masterpass-button-strategy.spec.ts
@@ -1,25 +1,26 @@
+
 import { createRequestSender, RequestSender } from '@bigcommerce/request-sender';
 import { createScriptLoader } from '@bigcommerce/script-loader';
 
-import { CheckoutButtonInitializeOptions } from '../';
-import { getCartState } from '../../cart/carts.mock';
+import { CheckoutButtonInitializeOptions } from '../..';
+import { getCartState } from '../../../cart/carts.mock';
 import {
     createCheckoutStore,
     Checkout,
     CheckoutActionCreator,
     CheckoutRequestSender,
     CheckoutStore
-} from '../../checkout';
-import { getCheckout, getCheckoutState } from '../../checkout/checkouts.mock';
-import { InvalidArgumentError, MissingDataError } from '../../common/error/errors';
-import { ConfigActionCreator, ConfigRequestSender } from '../../config';
-import { PaymentMethod } from '../../payment';
-import { getMasterpass, getPaymentMethodsState } from '../../payment/payment-methods.mock';
-import { Masterpass, MasterpassScriptLoader } from '../../payment/strategies/masterpass';
-import { getMasterpassScriptMock } from '../../payment/strategies/masterpass/masterpass.mock';
+} from '../../../checkout';
+import { getCheckout, getCheckoutState } from '../../../checkout/checkouts.mock';
+import { InvalidArgumentError, MissingDataError } from '../../../common/error/errors';
+import { ConfigActionCreator, ConfigRequestSender } from '../../../config';
+import { PaymentMethod } from '../../../payment';
+import { getMasterpass, getPaymentMethodsState } from '../../../payment/payment-methods.mock';
+import { Masterpass, MasterpassScriptLoader } from '../../../payment/strategies/masterpass';
+import { getMasterpassScriptMock } from '../../../payment/strategies/masterpass/masterpass.mock';
 
-import { CheckoutButtonStrategy, MasterpassButtonStrategy } from './';
-import { CheckoutButtonMethodType } from './checkout-button-method-type';
+import { CheckoutButtonStrategy, MasterpassButtonStrategy } from '..';
+import { CheckoutButtonMethodType } from '../checkout-button-method-type';
 
 describe('MasterpassCustomerStrategy', () => {
     let container: HTMLDivElement;
