+++ src/checkout-buttons/strategies/paypal/index.ts
@@ -1,0 +1,2 @@
+export { PaypalButtonInitializeOptions } from './paypal-button-options';
+export { default as PaypalButtonStrategy } from './paypal-button-strategy';
